+++ pages/_app.js
@@ -1,9 +1,13 @@
-import '../styles/globals.css'
+import "../styles/globals.css"
+
 import { SessionProvider } from "next-auth/react"
+
+import Header from '../components/header';
 
 function MyApp({ Component, pageProps }) {
   return (
     <SessionProvider>
+      <Header />
       <Component {...pageProps} />
     </SessionProvider>
   );