+++ CHANGELOG.md
@@ -17,3 +17,4 @@ All notable changes to this project will be documented in this file.
 - web-scalper api endpoints
 - Catamaran font family
 - Navigation component
+- Header component